+++ lib/buffer_image.dart
@@ -1,6 +1,8 @@
 library buffer_image;
 
+export 'src/abstract_image.dart';
 export 'src/buffer_image.dart';
+export 'src/gray_image.dart';
 export 'src/sample_mode.dart';
 export 'src/rgba_image.dart';
 export 'src/blend_mode.dart';