+++ lib/src/sample_mode.dart
@@ -1,10 +1,11 @@
 library sample_mode;
 
 import 'dart:math';
+import 'dart:ui';
 
 import 'package:flutter/painting.dart';
 
-import 'buffer_image.dart';
+import 'abstract_image.dart';
 
 /// abstract class of sample mode
 abstract class SampleMode {
@@ -16,10 +17,17 @@ abstract class SampleMode {
   final String mode;
   const SampleMode(this.mode);
 
-  Color sample(Point<double> point, BufferImage image, [Color? obColor]);
+  Color sample(Point<double> point, AbstractImage image, [Color? obColor]);
+
+  int sampleChannel(Point<double> point, AbstractImage image,
+      [ImageChannel? channel, int? obValue]);
 
   Color lerpColor(Color a, Color b, double t) {
     return Color.lerp(a, b, t)!;
+  }
+
+  int lerpChannel(int a, int b, double t) {
+    return lerpDouble(a, b, t)!.round();
   }
 }
 
@@ -29,8 +37,13 @@ class NearestSampleMode extends SampleMode {
 
   /// sample
   @override
-  Color sample(Point<double> point, BufferImage image, [Color? obColor]) {
+  Color sample(Point<double> point, AbstractImage image, [Color? obColor]) {
     return image.getColorSafe(point.x.round(), point.y.round(), obColor);
+  }
+
+  int sampleChannel(Point<double> point, AbstractImage image,
+      [ImageChannel? channel, int? obValue]) {
+    return image.getChannel(point.x.round(), point.y.round(), channel);
   }
 }
 
@@ -39,7 +52,7 @@ class BilinearSampleMode extends SampleMode {
   const BilinearSampleMode() : super('bilinear');
 
   @override
-  Color sample(Point<double> point, BufferImage image, [Color? obColor]) {
+  Color sample(Point<double> point, AbstractImage image, [Color? obColor]) {
     double x = point.x.floorToDouble();
     double y = point.y.floorToDouble();
 
@@ -47,13 +60,17 @@ class BilinearSampleMode extends SampleMode {
       if (y == point.y) {
         return image.getColorSafe(x.toInt(), y.toInt(), obColor);
       } else {
-        return lerpColor(image.getColorSafe(x.toInt(), point.y.floor(), obColor),
-            image.getColorSafe(x.toInt(), point.y.ceil(), obColor), point.y - y);
+        return lerpColor(
+            image.getColorSafe(x.toInt(), point.y.floor(), obColor),
+            image.getColorSafe(x.toInt(), point.y.ceil(), obColor),
+            point.y - y);
       }
     } else {
       if (y == point.y) {
-        return lerpColor(image.getColorSafe(point.x.floor(), y.toInt(), obColor),
-            image.getColorSafe(point.x.ceil(), y.toInt(), obColor), point.x - x);
+        return lerpColor(
+            image.getColorSafe(point.x.floor(), y.toInt(), obColor),
+            image.getColorSafe(point.x.ceil(), y.toInt(), obColor),
+            point.x - x);
       }
     }
 
@@ -71,6 +88,42 @@ class BilinearSampleMode extends SampleMode {
 
     return newColor;
   }
+
+  int sampleChannel(Point<double> point, AbstractImage image,
+      [ImageChannel? channel, int? obValue]) {
+    double x = point.x.floorToDouble();
+    double y = point.y.floorToDouble();
+
+    if (x == point.x) {
+      if (y == point.y) {
+        return image.getChannelSafe(x.toInt(), y.toInt(), obValue);
+      } else {
+        return lerpChannel(
+            image.getChannelSafe(x.toInt(), point.y.floor(), obValue),
+            image.getChannelSafe(x.toInt(), point.y.ceil(), obValue),
+            point.y - y);
+      }
+    } else {
+      if (y == point.y) {
+        return lerpChannel(
+            image.getChannelSafe(point.x.floor(), y.toInt(), obValue),
+            image.getChannelSafe(point.x.ceil(), y.toInt(), obValue),
+            point.x - x);
+      }
+    }
+
+    // tl, tr, br, bl
+    List<int> colors = [
+      image.getChannelSafe(point.x.floor(), point.y.floor(), obValue),
+      image.getChannelSafe(point.x.ceil(), point.y.floor(), obValue),
+      image.getChannelSafe(point.x.ceil(), point.y.ceil(), obValue),
+      image.getChannelSafe(point.x.floor(), point.y.ceil(), obValue)
+    ];
+    double tHor = point.x - x, tVer = point.y - y;
+
+    return lerpChannel(lerpChannel(colors[0], colors[1], tHor),
+        lerpChannel(colors[3], colors[2], tHor), tVer);
+  }
 }
 
 /// Bicubic sample mode @Unimplemented
@@ -78,8 +131,15 @@ class BicubicSampleMode extends SampleMode {
   const BicubicSampleMode() : super('bicubic');
 
   @override
-  Color sample(Point<double> point, BufferImage image, [Color? obColor]) {
+  Color sample(Point<double> point, AbstractImage image, [Color? obColor]) {
     // TODO: implement sample
+    throw UnimplementedError();
+  }
+
+  @override
+  int sampleChannel(Point<double> point, AbstractImage image,
+      [ImageChannel? channel, int? obValue]) {
+    // TODO: implement sampleChannel
     throw UnimplementedError();
   }
 }
@@ -89,8 +149,15 @@ class LanczosSampleMode extends SampleMode {
   const LanczosSampleMode() : super('lanczos');
 
   @override
-  Color sample(Point<double> point, BufferImage image, [Color? obColor]) {
+  Color sample(Point<double> point, AbstractImage image, [Color? obColor]) {
     // TODO: implement sample
+    throw UnimplementedError();
+  }
+
+  @override
+  int sampleChannel(Point<double> point, AbstractImage image,
+      [ImageChannel? channel, int? obValue]) {
+    // TODO: implement sampleChannel
     throw UnimplementedError();
   }
 }